backend/requirements.txt
flask
flask-cors
spacy
pdfplumber
python-docx
werkzeug
groq
gunicorn